Begin extracting helper functions for consolidation
