+++ iati/tests/base_functional_tests.py
@@ -2,7 +2,7 @@
 import os
 import string
 import random
-import time
+# import time
 import pytest
 from conftest import LOCALHOST
 from django.core.management import call_command
@@ -24,42 +24,42 @@ DEFAULT_PAGES = DEFAULT_PAGES + [{'title': 'Home', 'slug': '', 'model': HomePage
 TEST_DATA_DIR = settings.BASE_DIR + '/tests/data/'
 
 
-def view_live_page(admin_browser, page_title):
-    """Navigate to the published page on the site.
+# def view_live_page(admin_browser, page_title):
+#     """Navigate to the published page on the site.
+#
+#     Args:
+#         page_title (str): The page title text you are expecting on the live page.
+#
+#     """
+#     button_links = admin_browser.find_by_xpath('//a[@title="View live version of \'{}\'"]'.format(page_title))
+#     href = button_links[0].__dict__['_element'].get_property('href')
+#     admin_browser.visit(href)
+#
+#
+# def prevent_alerts(admin_browser):
+#     """Stop the Wagtail CMS from sending beforeunload alerts.
+#
+#     Args:
+#         admin_browser (browser): The splinter browser instance.
+#
+#     """
+#     admin_browser.driver.execute_script("window.removeEventListener('beforeunload', window.areYouSure);")
 
-    Args:
-        page_title (str): The page title text you are expecting on the live page.
 
-    """
-    button_links = admin_browser.find_by_xpath('//a[@title="View live version of \'{}\'"]'.format(page_title))
-    href = button_links[0].__dict__['_element'].get_property('href')
-    admin_browser.visit(href)
-
-
-def prevent_alerts(admin_browser):
-    """Stop the Wagtail CMS from sending beforeunload alerts.
-
-    Args:
-        admin_browser (browser): The splinter browser instance.
-
-    """
-    admin_browser.driver.execute_script("window.removeEventListener('beforeunload', window.areYouSure);")
-
-
-def wait_for_clickability(element, wait_time=1):
-    """Wait until an element is enabled before clicking.
-
-    Args:
-        element (ElementAPI): The splinter element to be waited on.
-        wait_time (int): The time in seconds to wait.
-
-    """
-    end_time = time.time() + wait_time
-
-    while time.time() < end_time:
-        if element and element.__dict__['_element'].is_enabled():
-            return True
-    return False
+# def wait_for_clickability(element, wait_time=1):
+#     """Wait until an element is enabled before clicking.
+#
+#     Args:
+#         element (ElementAPI): The splinter element to be waited on.
+#         wait_time (int): The time in seconds to wait.
+#
+#     """
+#     end_time = time.time() + wait_time
+#
+#     while time.time() < end_time:
+#         if element and element.__dict__['_element'].is_enabled():
+#             return True
+#     return False
 
 
 def wait_for_visibility(element, wait_time=1):
@@ -103,16 +103,16 @@ def random_string(size=10, chars=string.ascii_uppercase + string.ascii_lowercase
     return ''.join(random.choice(chars) for _ in range(size))
 
 
-def click_obscured(admin_browser, element):
-    """A function that clicks elements even if they're slightly obscured.
-
-    Args:
-        admin_browser (browser): The splinter browser instance.
-        element (ElementAPI): The splinter element to be waited on.
-
-    """
-    wait_for_clickability(element)
-    admin_browser.driver.execute_script("arguments[0].click();", element.__dict__['_element'])
+# def click_obscured(admin_browser, element):
+#     """A function that clicks elements even if they're slightly obscured.
+#
+#     Args:
+#         admin_browser (browser): The splinter browser instance.
+#         element (ElementAPI): The splinter element to be waited on.
+#
+#     """
+#     wait_for_clickability(element)
+#     admin_browser.driver.execute_script("arguments[0].click();", element.__dict__['_element'])
 
 
 def scroll_to_element(admin_browser, element):
