+++ iati/tests/iati_standard_functional_tests.py
@@ -1,33 +1,7 @@
 """A module of functional tests for the IATI Standard page."""
 import pytest
-from tests.base_functional_tests import click_obscured
-from tests.about_functional_tests import reveal_content_editor, scroll_to_bottom_of_page, TEST_DATA_DIR, view_live_page
-
-
-def navigate_to_Home_cms_section(admin_browser):
-    """Navigate to the Home section of the CMS."""
-    admin_browser.click_link_by_text('Pages')
-    admin_browser.find_by_text('Home').click()
-
-
-def upload_an_image(admin_browser):
-    """Upload an image in the CMS.
-
-    Note:
-        Duplicate of the same helper function in base_functional_tests.
-
-    """
-    admin_browser.find_by_text('Choose an image').click()
-    click_obscured(admin_browser, admin_browser.find_by_text('Upload').first)
-    admin_browser.fill('title', 'Test image')
-    admin_browser.attach_file('file', TEST_DATA_DIR + 'pigeons.jpeg')
-    admin_browser.find_by_xpath('//em[contains(text(), "Upload")]').click()
-
-
-def publish_changes(admin_browser):
-    """Publish changes made in the CMS to the live page."""
-    click_obscured(admin_browser, admin_browser.find_by_xpath('//div[@class="dropdown-toggle icon icon-arrow-up"]').first)
-    click_obscured(admin_browser, admin_browser.find_by_text('Publish').first)
+from tests import helper_functions
+from tests.about_functional_tests import reveal_content_editor, scroll_to_bottom_of_page
 
 
 @pytest.mark.django_db
@@ -37,29 +11,29 @@ class TestIATIStandardPageisEditable():
 # I want to be able to edit the header text of this page
     def test_header_can_be_edited(self, admin_browser):
         """Check that the page header can be edited in the CMS."""
-        navigate_to_Home_cms_section(admin_browser)
+        helper_functions.navigate_to_Home_cms_section(admin_browser)
         admin_browser.find_by_text('IATI Standard').click()
         admin_browser.find_by_text('English')[0].click()
         admin_browser.fill('heading_en', 'IATI Standard')
-        publish_changes(admin_browser)
-        view_live_page(admin_browser, 'IATI Standard')
+        helper_functions.publish_changes(admin_browser)
+        helper_functions.view_live_page(admin_browser, 'IATI Standard')
         assert admin_browser.is_text_present('IATI Standard')
 
 # I want to be able to edit the excerpt of this page
     def test_excerpt_can_be_edited(self, admin_browser):
         """Check that the page excerpt can be edited in the CMS."""
-        navigate_to_Home_cms_section(admin_browser)
+        helper_functions.navigate_to_Home_cms_section(admin_browser)
         admin_browser.find_by_text('IATI Standard').click()
         admin_browser.find_by_text('English')[0].click()
         admin_browser.fill('excerpt_en', 'This is an excerpt.')
-        publish_changes(admin_browser)
-        view_live_page(admin_browser, 'IATI Standard')
+        helper_functions.publish_changes(admin_browser)
+        helper_functions.view_live_page(admin_browser, 'IATI Standard')
         assert admin_browser.is_text_present('This is an excerpt.')
 
 # I want to be able to add summary content to this page
     def test_body_content_can_be_edited_simple(self, admin_browser):
         """Check that basic text content from the content editor in the CMS shows on the page."""
-        navigate_to_Home_cms_section(admin_browser)
+        helper_functions.navigate_to_Home_cms_section(admin_browser)
         admin_browser.find_by_text('IATI Standard').click()
         admin_browser.find_by_text('English')[0].click()
         element_count = admin_browser.find_by_id('content_editor_en-count').value
@@ -67,8 +41,8 @@ class TestIATIStandardPageisEditable():
         reveal_content_editor(admin_browser, 'Intro', element_count)
         admin_browser.find_by_text('Intro')[int(element_count)].click()
         admin_browser.find_by_xpath('//div[@class="notranslate public-DraftEditor-content"]').fill('This is some content.')
-        publish_changes(admin_browser)
-        view_live_page(admin_browser, 'IATI Standard')
+        helper_functions.publish_changes(admin_browser)
+        helper_functions.view_live_page(admin_browser, 'IATI Standard')
         assert admin_browser.is_text_present('This is some content.')
 
 
@@ -81,7 +55,7 @@ class TestRedirectLinksWorking():
 
     def visit_iati_standard_page(self, admin_browser):
         """Go to the IATI Stadard page from the admin site."""
-        navigate_to_Home_cms_section(admin_browser)
+        helper_functions.navigate_to_Home_cms_section(admin_browser)
         page_button = admin_browser.find_by_xpath('//a[@href="/en/iati-standard/"]').first
         page_link = page_button._element.get_property('href')
         self.LOCALHOST_LINK = page_link[:-15]
